requirements.txt
# Dependencies
streamlit==1.31.0
pandas>=2.0.0
python-dotenv==1.0.0
openai==1.12.0
google-api-python-client==2.100.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
youtube-transcript-api==0.6.1
python-dateutil==2.8.2
werkzeug==3.0.1
httplib2>=0.22.0

# Python version constraint
python_version>="3.11.7" 